backend/chemdb/requirements.txt
You need:

Django
djangorestframework
psycopg2-binary  # For PostgreSQL
